requirements.txt
google-adk>=0.1.0
google-genai>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
